[#67] chore: ErrorHandling 폴더링
